Move Tuya type information classes to separate module
