+++ homeassistant/components/tuya/light.py
@@ -35,8 +35,8 @@ from .models import (
     DPCodeEnumWrapper,
     DPCodeIntegerWrapper,
     DPCodeJsonWrapper,
-    IntegerTypeData,
 )
+from .type_information import IntegerTypeInformation
 from .util import remap_value
 
 
@@ -138,24 +138,24 @@ class _ColorTempWrapper(DPCodeIntegerWrapper):
         )
 
 
-DEFAULT_H_TYPE = IntegerTypeData(
+DEFAULT_H_TYPE = IntegerTypeInformation(
     dpcode=DPCode.COLOUR_DATA_HSV, min=1, scale=0, max=360, step=1
 )
-DEFAULT_S_TYPE = IntegerTypeData(
+DEFAULT_S_TYPE = IntegerTypeInformation(
     dpcode=DPCode.COLOUR_DATA_HSV, min=1, scale=0, max=255, step=1
 )
-DEFAULT_V_TYPE = IntegerTypeData(
+DEFAULT_V_TYPE = IntegerTypeInformation(
     dpcode=DPCode.COLOUR_DATA_HSV, min=1, scale=0, max=255, step=1
 )
 
 
-DEFAULT_H_TYPE_V2 = IntegerTypeData(
+DEFAULT_H_TYPE_V2 = IntegerTypeInformation(
     dpcode=DPCode.COLOUR_DATA_HSV, min=1, scale=0, max=360, step=1
 )
-DEFAULT_S_TYPE_V2 = IntegerTypeData(
+DEFAULT_S_TYPE_V2 = IntegerTypeInformation(
     dpcode=DPCode.COLOUR_DATA_HSV, min=1, scale=0, max=1000, step=1
 )
-DEFAULT_V_TYPE_V2 = IntegerTypeData(
+DEFAULT_V_TYPE_V2 = IntegerTypeInformation(
     dpcode=DPCode.COLOUR_DATA_HSV, min=1, scale=0, max=1000, step=1
 )
 
@@ -578,15 +578,15 @@ def _get_color_data_wrapper(
     if function_data := json_loads_object(
         cast(str, color_data_wrapper.type_information.type_data)
     ):
-        color_data_wrapper.h_type = IntegerTypeData(
+        color_data_wrapper.h_type = IntegerTypeInformation(
             dpcode=color_data_wrapper.dpcode,
             **cast(dict, function_data["h"]),
         )
-        color_data_wrapper.s_type = IntegerTypeData(
+        color_data_wrapper.s_type = IntegerTypeInformation(
             dpcode=color_data_wrapper.dpcode,
             **cast(dict, function_data["s"]),
         )
-        color_data_wrapper.v_type = IntegerTypeData(
+        color_data_wrapper.v_type = IntegerTypeInformation(
             dpcode=color_data_wrapper.dpcode,
             **cast(dict, function_data["v"]),
         )
